+++ src/Components/Card.php
@@ -83,7 +83,7 @@ class Card extends AbstractComponent {
 			'class' => $this->arrayToString( $innerClass, ' ' ),
 		];
 		if ( $this->isCollapsible() ) {
-			$innerAttributes['data-parent'] = $this->getDataParent();
+			$innerAttributes['data-bs-parent'] = $this->getDataParent();
 			$innerAttributes['aria-labelledby'] = $this->getId() . '_header';
 		}
 
@@ -240,8 +240,8 @@ class Card extends AbstractComponent {
 		if ( $type == 'header' ) {
 			if ( $this->isCollapsible() ) {
 				$newAttributes += [
-					'data-toggle'   => 'collapse',
-					'data-target'   => '#' . $this->getId(),
+					'data-bs-toggle'   => 'collapse',
+					'data-bs-target'   => '#' . $this->getId(),
 					'aria-controls' => $this->getId(),
 					'aria-expanded' => $this->getValueFor( 'active' ) ? 'true' : 'false',
 					'id'            => $this->getId() . '_header'